Solution structure created. Added projects "Core", "BussnesObject". Links restored.
BaseElement at work.
