+++ BussinesObject/PageObjects/LoginPage.cs
@@ -1,34 +1,44 @@
-﻿using Diploma.PageObject;
+﻿using Diploma.Core;
 using OpenQA.Selenium;
-using SeleniumTests.Diploma.PageObject;
 using SeleniumTests.Diploma;
 using SeleniumTests.Diploma.PageObject;
-using Diploma.Core;
 
 namespace Diploma.PageObject
 {
-    public class AuthenticationPage : BasePage
+    public class LoginPage : BasePage
     {
         private By UserMailInput = By.Id("email");
         private By PasswordInput = By.Id("passwd");
-        //private By ErrorMessage = By.ClassName("alert alert-danger");
+        private By ErrorMessage = By.XPath("//*[@class='alert alert-danger'][.//li[contains(text(),'Authentication failed')]]");
         private By SubmitLoginButtom = By.Id("SubmitLogin");
+        private By LoginButtom = By.ClassName("login");
+
+        public const string url = "http://prestashop.qatestlab.com.ua/ru/authentication?back=my-account";
 
 
-        public const string url =
-        "http://prestashop.qatestlab.com.ua/ru/authentication?multi-shipping=0&display_guest_checkout=0&back=http%3A%2F%2Fprestashop.qatestlab.com.ua%2Fru%2Forder%3Fstep%3D1%26multi-shipping%3D0";
+        public LoginPage()
+        {
+            WaitHelper.WaitElement(driver, LoginButtom);
+        }
 
-        public override AuthenticationPage OpenPage()
+        public override LoginPage OpenPage()
         {
             Browser.Instance.NavigateToUrl(url);
             return this;
         }
+        
+        public AccountPage LoginAsStandartUser()
+        {
+            var user = UserBuilder.GetStandandartUser();  
+            TryToLogin(user);
 
+            return new AccountPage();
+        }
         public AddressPage LoginAndGoToAdressPage()
         {
             var user = UserBuilder.GetStandandartUser();
-            
             TryToLogin(user);
+
             return new AddressPage();
         }
 
@@ -37,6 +47,14 @@ namespace Diploma.PageObject
             driver.FindElement(UserMailInput).SendKeys(user.Mail);
             driver.FindElement(PasswordInput).SendKeys(user.Password);
             driver.FindElement(SubmitLoginButtom).Click();
+        }
+
+
+
+        public bool VerifyErrorMesage()
+        {
+            driver.FindElement(ErrorMessage);
+            return false;
         }
     }
 }
