+++ Core/BaseElement.cs
