+++ Diploma.sln
@@ -3,7 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33723.286
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Diploma", "Diploma\Diploma.csproj", "{8504DA3F-12D6-4E37-ADD2-B04F7859AF81}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Diploma\Tests.csproj", "{8504DA3F-12D6-4E37-ADD2-B04F7859AF81}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{29DCE8C4-8891-40DA-A921-516CFD0CBF5D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BussinesObject", "BussinesObject\BussinesObject.csproj", "{E004EC2B-E8F5-4943-A5CA-B90B7D8B10FF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{8504DA3F-12D6-4E37-ADD2-B04F7859AF81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8504DA3F-12D6-4E37-ADD2-B04F7859AF81}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8504DA3F-12D6-4E37-ADD2-B04F7859AF81}.Release|Any CPU.Build.0 = Release|Any CPU
+		{29DCE8C4-8891-40DA-A921-516CFD0CBF5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{29DCE8C4-8891-40DA-A921-516CFD0CBF5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{29DCE8C4-8891-40DA-A921-516CFD0CBF5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{29DCE8C4-8891-40DA-A921-516CFD0CBF5D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E004EC2B-E8F5-4943-A5CA-B90B7D8B10FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E004EC2B-E8F5-4943-A5CA-B90B7D8B10FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E004EC2B-E8F5-4943-A5CA-B90B7D8B10FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E004EC2B-E8F5-4943-A5CA-B90B7D8B10FF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
