+++ Diploma/RunSettings/BrowserSetup.runsettings
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <RunSettings>
 	<TestRunParameters>
-		<Parameter name="Headless" value="true"/>
+		<Parameter name="Headless" value="false"/>
 		<Parameter name="BrowserType" value="Chrome"/>
 		<Parameter name="ImplicityWait" value="10"/>
